Silent IBD salvage during dojo uninstall

+++ Scripts/Install/install-dojo.sh
@@ -309,8 +309,11 @@ EOF
 
         _sleep
 
-        sudo rm -rf "${DOCKER_VOLUME_BITCOIND}"/_data/{blocks,chainstate}
-        sudo mv -v "${INSTALL_DIR_UNINSTALL}"/{blocks,chainstate} "${DOCKER_VOLUME_BITCOIND}"/_data/ 1>/dev/null
+        if sudo test -d "${DOCKER_VOLUME_BITCOIND}"/_data/blocks; then
+            sudo rm -rf "${DOCKER_VOLUME_BITCOIND}"/_data/{blocks,chainstate}
+        fi
+
+        sudo mv "${INSTALL_DIR_UNINSTALL}"/{blocks,chainstate} "${DOCKER_VOLUME_BITCOIND}"/_data/ 1>/dev/null
         # changes to dojo path, otherwise exit
         # websearch "bash Logical OR (||)" for info
         # stops dojo and removes new data directories
